dont delete alliance if users remain

+++ src/Component/Player/Deletion/Handler/AllianceDeletionHandler.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Stu\Component\Player\Deletion\Handler;
 
 use Override;
+use Doctrine\Common\Collections\Collection;
 use Stu\Component\Alliance\AllianceEnum;
 use Stu\Module\Alliance\Lib\AllianceActionManagerInterface;
 use Stu\Orm\Entity\UserInterface;
@@ -12,9 +13,7 @@ use Stu\Orm\Repository\AllianceJobRepositoryInterface;
 
 final class AllianceDeletionHandler implements PlayerDeletionHandlerInterface
 {
-    public function __construct(private AllianceJobRepositoryInterface $allianceJobRepository, private AllianceActionManagerInterface $allianceActionManager)
-    {
-    }
+    public function __construct(private AllianceJobRepositoryInterface $allianceJobRepository, private AllianceActionManagerInterface $allianceActionManager) {}
 
     #[Override]
     public function delete(UserInterface $user): void
@@ -25,10 +24,18 @@ final class AllianceDeletionHandler implements PlayerDeletionHandlerInterface
 
                 $successor = $alliance->getSuccessor();
 
-                if ($successor === null) {
+                $diplomatic = $alliance->getDiplomatic();
+
+                $members = $alliance->getMembers();
+                $members->removeElement($user);
+
+                $lastonlinemember = $this->getLastOnlineMember($members);
+
+                if ($successor === null && $lastonlinemember === null) {
                     $this->allianceJobRepository->delete($job);
-                    $this->allianceActionManager->delete($alliance->getId(), false);
-                } else {
+                    $this->allianceActionManager->delete($alliance->getId(), true);
+                }
+                if ($successor !== null) {
                     $successorUserId = $successor->getUserId();
 
                     $this->allianceActionManager->setJobForUser(
@@ -38,9 +45,39 @@ final class AllianceDeletionHandler implements PlayerDeletionHandlerInterface
                     );
                     $this->allianceJobRepository->delete($successor);
                 }
+                if ($successor == null && $lastonlinemember != null) {
+                    if ($diplomatic !== null) {
+                        if ($lastonlinemember == $diplomatic->getUser()) {
+                            $this->allianceJobRepository->delete($diplomatic);
+                        }
+                    }
+                    $this->allianceActionManager->setJobForUser(
+                        $alliance->getId(),
+                        $lastonlinemember->getId(),
+                        AllianceEnum::ALLIANCE_JOBS_FOUNDER
+                    );
+                }
             } else {
                 $this->allianceJobRepository->delete($job);
             }
         }
+    }
+
+    /**
+     * @param Collection<int, UserInterface> $members
+     */
+    private function getLastOnlineMember(Collection $members): ?UserInterface
+    {
+        $lastOnlineMember = null;
+        $maxLastAction = 0;
+
+        foreach ($members as $member) {
+            if ($member->getLastAction() > $maxLastAction) {
+                $maxLastAction = $member->getLastAction();
+                $lastOnlineMember = $member;
+            }
+        }
+
+        return $lastOnlineMember;
     }
 }
